Feature/junit

+++ timeskip-ejb/src/main/java/be/ehb/facades/ManagementFacade.java
@@ -52,7 +52,7 @@ public class ManagementFacade implements IManagementFacade {
         MembershipBean newMembership = new MembershipBean();
         newMembership.setRoleId(roleId);
         newMembership.setUserId(userId);
-        newMembership.setRoleId(roleId);
+        newMembership.setOrganizationId(organizationId);
         return ResponseFactory.createMembershipResponse(storage.createOrUpdateMembership(newMembership));
     }
 
@@ -170,6 +170,8 @@ public class ManagementFacade implements IManagementFacade {
     @Override
     public void deletePaygrade(Long paygradeId) {
         PaygradeBean paygrade = storage.getPaygrade(paygradeId);
+        if (!storage.findUsersByPaygrade(paygradeId).isEmpty())
+            throw ExceptionFactory.paygradeStillInUseException(paygradeId);
         storage.deletePaygrade(paygrade);
     }
 }
